feat: add calculate double and cube
add the calculate double and cube feature to the project. User provides two numbers
only and we determine the double and cube for both numbers displaying the doubles for
both numbers then the cube for both numbers seperately.

Had to put a limitation on the user input to just 2 becasue things are done pretty manually
now.

Resolves: #4